Handle flings even if not in control
Our current nested scrolling implementation relies on
onNested[Pre]Scroll events to start a WindowInsetsAnimation control
request. There's a scenario though where we do not receive any scroll
events, but we do receive an onNested[Pre]Fling. This can happen if the
a parent is consuming, or touch slop isn't reached, etc.

Currently that fling request would be dropped, because our controller
would not be in control of the IME. This commit fixes that situation by
starting a control request, followed by an immediate fling once in
control. A new `startAndFling` function has been added for this.

Also increased the number of messages to be able to fling for longer
distances in the sample.

Change-Id: Id83937ad2283c2d4c05458dc4ff8ab6277b5ffcb