Add a AppWidget sample
At this moment this sample includes widget with a grocery list with checkboxes
that demonstrates:
- Description in widget preview
- Preview layout
- Dynamic coloring
- Rounded corners and padding

Other APIs are going to be added in follow up CLs.
It may be ideal to wire the grocery list with SQLite database, but it's
not prioritized because that part is not relevant with the Widget APIs.
(PS1 only adds an Empty Activity to make the review easier)

Change-Id: Ic97986f1716221a6338599aad6f8cba43a80d0d7
(cherry picked from commit 341d2bb29311b23e2a85564e4941f4cbba437f25)