Update settings.gradle to work with Android Studio Arctic Fox
Before this change, the Android Gradle Plugin version had been updated to 7.1, which is only compatible with Android Studio Bumblebee and Chipmunk, which are both currently only in preview. Downgrading to 7.0 makes the sample compatible with Android Studio Arctic Fox, which is the current stable release.

This commit also removes some unused configuration.